Update patch set 3
Patch Set 3: Commit-Queue+1

> Patch Set 3: Code-Review+1
> 
> I guess this means a GCM suite won't actually be selected unless you do SDP munging? Should SelectCrypto maybe be changed to prefer GCM? That gets us halfway there...

The goal of GCM in general is to make it cheaper for middleboxes to do crypto, so we're delivering value even if we're not putting GCM first (as middleboxes can choose to only offer GCM). 

GCM does have a longer auth tag, and thus more overhead, so moving to prefer GCM for chrome-chrome calls is an entirely separate discussion that can occur later.

Patch-set: 3
Reviewer: Gerrit User 8393 <8393@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+1