Update patch set 2
Patch Set 2:

PTAL. It looks like the fWait_ variable is used as an exit condition in the event loops, and set by the PhysicalSocketServer::WakeUp method, with a bit of indirection. To me, it seems some kind of synchronization would be appropriate, but not touched in this cl.

Patch-set: 2
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0
Work-in-progress: false