Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

Hmm.... this is a coordination problem between the ping frequency and the timeout value.

Consider entities A and B, default timeout.

If A then sends a ping every 10 sec, all should be good (most of the time).
If A loses 2 consecutive pings, there will still be incoming traffic at B after 30 sec, and we'll survive (barely). If it loses 3, we're dead.

If B has a longer timeout (60 s), it will survive more packet losses, and A can send less frequently.

We should also be able to timeout faster when we have an outstanding ping; 30 secs seems like a long time for this. But perhaps this is checked elsewhere?

Changing the timing rules on one side always scares me a little. But lengthening the timeout shouldn't be the part that's harmful (if any).

Note: I'm assuming that incoming pings and incoming ping responses both update last_received(); I haven't verified that this is true.

Patch-set: 3
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1