Update patch set 7
Patch Set 7:

(1 comment)

> Patch Set 5:
> 
> lgtm, but what's the purpose of scalable_video_controller_no_layering.{h,cc}?

It is a simple placeholder controller for no scalability.
One of the alternatives is to pass nullptr and have all no scalability logic inside the av1 encoder,
but I prefer to keep it externally. In particular that allows to turn scalability on and off at runtime.

Patch-set: 7
Reviewer: Gerrit User 5019 <5019@58829da1-049c-39fb-b951-ebdcd0984767>