Update patch set 2
Patch Set 2:

How long will it take for the fallback to happen if the decoder is actually broken? If it fails to decode, then we have to wait for a keyframe until we try and fail again. If you need 4 consecutive fails for this to happen then it could become quite noticeable.

Patch-set: 2
CC: Gerrit User 5255 <5255@58829da1-049c-39fb-b951-ebdcd0984767>