Update patch set 2
Patch Set 2: Code-Review-1

I personally think this is a bad idea.

It increases variability of platform behavior across the Web, meaning that applications will need to have multiple strategies to deal with multiple configurations - and the fact that they have no way of knowing which particular policies are in effect means that they will either have to stick with what's permitted under all policies, or probe for what can be permitted under any policy.

The concerns raised in the bug (which needs to be linked in the description, btw) are, to my mind, not sufficient to warrant the impact on interoperability.

We should take the discussion to the bug.

Patch-set: 2
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=-1