Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2: Code-Review-1
> 
> I personally think this is a bad idea.
> 
> It increases variability of platform behavior across the Web, meaning that applications will need to have multiple strategies to deal with multiple configurations - and the fact that they have no way of knowing which particular policies are in effect means that they will either have to stick with what's permitted under all policies, or probe for what can be permitted under any policy.
> 
> The concerns raised in the bug (which needs to be linked in the description, btw) are, to my mind, not sufficient to warrant the impact on interoperability.

The default here is to have mDNS enabled, which would keep alive interop in pretty much all cases.  DNS (non-mDNS) candidates have never really been a thing (and didn't exist in standalone WebRTC until January of this year, nor in Chrome until last year), so I can't see how turning those off is going to cause anyone any trouble.

But I can understand if you want to remove the knob for turning off the mDNS hostname lookups since it's really just for the very-privacy-sensitive, but they can just fork WebRTC if they have to.

And I can understand if you want to remove the knob for turning on the DNS hostname lookups since, as mentioned, DNS (non-DNS) candidates have never really been a thing.

So I suppose we could change this entire CL into "only do mDNS hostname lookups, not DNS hostname lookups" and make that behavior not just the default but the only behavior with configurability.  That's how I had it originally, actually, but then I thought you might want to keep the DNS lookups for Chrome (for some reason, just in case, even though the likelyhood of breaking anything is slim).  And then it seemed that if there was 1 knob there might as well be 2.

But I think it's wrong to just leave DNS lookups on because it provides little to no value and a lot of users of standalone WebRTC will unknowingly be caught in the issues it causes.

Patch-set: 2