Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Overall looks good, I'll have a closer look this afternoon.
> A general question though. Is it worth spending time optimizing for ProcessThread or should we just push to get rid of it asap? Most places seem like they could trivially run on task queues instead....

Depends on what 'asap' means. I took a look at the pacer changes and it looks like the work has been going on for a while. This CL is a way to make some progress now and not "not do something" because something else might happen sometime soon 😊

Fully getting rid of the process thread is an ongoing task as far as I'm aware, but due to how it's being used by various different projects that use webrtc, it's still not risk free or simple.

Patch-set: 4
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0