Update patch set 1
Patch Set 1:

Heh. The point of that log was to be obnoxious, because turning off encryption is a very singificant operator error in all cases except when testing that you can turn it off.

Unfortunately RTC_LOG doesn't have the equivalent of LOG_ONCE, which is a pity, because it would have been my preferred solution to this problem - it's an important message to give, but we don't really need to say it a thousand times.

Patch-set: 1
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>