Update patch set 5
Patch Set 5: Commit-Queue+1

Hello Tommi,

Could you please have a look at this CL which makes error messages more clear when SDP application is failed. Error message will now say in which m-section something went wrong. This is valuable in conferencing scenarios when there are many m-sections due to many peers in a conference. When something is going wrong it is really valuable to have a error message which narrow down to specific m-section. I'm not sure it is not breaking some tests in Chromium if they rely on particular error message, but it is always possible to revert this CL if it happen.

Thanks a lot in advance!

Patch-set: 5
Reviewer: Gerrit User 6305 <6305@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+1
Work-in-progress: false