Update patch set 2
Patch Set 2:

Hello Olga,

Thanks for letting know that I've broken tests in Chromium project.

I'm curious what is correct way to get original change merged and not to break tests?

I can't update existing tests in Chromium in advance because it will also break them and I can't update WebRTC. So there should be a way to create "linked" CLs in Chromium and WebRTC which could only be accepted as a whole. Is that right?

Thanks a lot in advance!

Patch-set: 2
Reviewer: Gerrit User 6305 <6305@58829da1-049c-39fb-b951-ebdcd0984767>