Update patch set 4
Patch Set 4:

Karl PTAL.

See PRESUBMIT.py for the check and the docs added to https://webrtc.googlesource.com/src/+/refs/heads/master/abseil-in-webrtc.md.

Some bots are failing but there are other CLs out to take care of them.

The package absl/flags is testonly, hence not part of the component, this is why I allow direct deps on it. The package absl/debugging has a static initializer in it so I allow a static dependency on it until it will be part of the component.

Patch-set: 4
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5087 <5087@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false