Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review+1
> > 
> > OK, then. Opus made a wrong choice in not exposing the number, we'll just live with it.
> 
> Actually, I think it might be fishy to rely on this number, as it is implementational specific and does not seem to be defined in the Opus spec.

Opus is a bit unusual in that they chose to make the C code normative - if it behaves like the code in the RFC, it's correct. So depending on implementation details is fine - as long as it's implementation details of the RFC version of the code.

Makes for a bit of trouble when improving the code - you can improve performance, but you can never improve behavior, since the correct behavior is defined by the code.

So we're *probably* safe. Until someone starts making "improved" Opus versions that break spec :-(

Patch-set: 3
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>