Update patch set 2
Patch Set 2: Code-Review+1

Approved as-is but I do suspect that keeping the old name around for one milestone might be less disruptive but I'll leave that to you. And I bet there are other obsolete renames we could delete too.

Up to you. And FYI I'll be OOO now, back on Tuesday, so if you need a stamp on the chromium whielist, ask guidou for revieew

Patch-set: 2
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1