Update patch set 10
Patch Set 10:

(9 comments)

Wow, this is a lot of stuff. :) 


See some more comments. One thing I am a bit worried is that I think we are copying too many strings in the OnSomething functions. It might be premature optimization, but I have the feeling that we should do better. I mean, based on how many frames we receive in a test we might end up creating the same StatsKey over and over and copying strings in the process. Maybe using string_view and hashing is a possible way to avoid this?

Maybe we should have a meeting to convince me this is not a problem for the video analyzer pipeline.

Patch-set: 10