Update patch set 11
Patch Set 11:

thakis@ for the DEPS change to depend on //testing/gtest. We use absl/flags for tests and WebRTC testonly code should not be use in Chromium, except for this target, which is used to test the TaskQueue integration in Chromium. Switching to //testing/gtest allows us to cut the dependency path to absl/flags (which is not part of the Abseil component).

+kwiberg@ for everything else.

Patch-set: 11
Reviewer: Gerrit User 5509 <5509@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false