Update patch set 1
Patch Set 1:

The point of this class is to enforce lifetime and correct access to the OES texture. If you send out multiple frames referencing the same OES texture, this breaks. I.e. if all references to the first frames are released, we will call returnTextureFrame(), even though a second frame might still be referencing this texture. This is probably possible to fix with some further logic. However, I would really prefer if you figure out a way to have this logic outside of SurfaceTextureHelper in a component that is more encapsulated and related to the problem. This could be in either the ScreenCapturer or in the encoder if the encoder is requiring a minimum frame rate.

Patch-set: 1