Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The point of this class is to enforce lifetime and correct access to the OES texture. If you send out multiple frames referencing the same OES texture, this breaks. I.e. if all references to the first frames are released, we will call returnTextureFrame(), even though a second frame might still be referencing this texture. This is probably possible to fix with some further logic. However, I would really prefer if you figure out a way to have this logic outside of SurfaceTextureHelper in a component that is more encapsulated and related to the problem. This could be in either the ScreenCapturer or in the encoder if the encoder is requiring a minimum frame rate.

I don't think this generates a new frame before the previous has been released. Based on the my read
of the code, I think it looks like it would work.

But I agree it doesn't really feel like it belongs here. And I am a bit scared to but more logic in
this class given it doesn't have great testing behind it.

However, I see why this would be a convenient place to put the logic. Because anywhere else it is
not really possible to get a new reference to a past frame.

One thing to consider is that there is nothing technically forcing using this specific implementation,
one could implement their own "SurfaceTextureHelper" specific to the capturer implementation.

If we end up making a change here, I think my preference would be the minimum possible, which I think
is just adding a method for forcing out a new frame. I am not even sure not calling updateTexImage()
is needed.

All of the rest (timestamp manipulation, poll loop etc.) can potentially be done outside.

Patch-set: 1