Update patch set 6
Patch Set 6:

PTAL. This cl deletes the remains of the MediaTransportInterface integration.

The DataChannelTransportInterface is left as is. That seems to be a clean and independent interface, even though it's not entirely clear to me if it is in use.

Patch-set: 6
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5525 <5525@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5845 <5845@58829da1-049c-39fb-b951-ebdcd0984767>
CC: Gerrit User 7565 <7565@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false