[Adaptation] Reset video restrictions in VideoStreamAdapter
This moves the functionality from ResourceAdaptationProcessor
to the VideoStreamAdapter. The ResourceAdaptationProcessor now
listens for the reset event and performs its needed functionality.

This means that going forward VideoStreamAdapter can be reset in
scenarios like switching sources, instead of having to do it in the
ResourceAdaptationProcessor. Additionally, this may happen on a
different queue than the ResourceAdaptationProcessor.