Update patch set 3
Patch Set 3:

Hi Taylor - I took a closer look at the DataChannel implementation (and DataChannelController) and see that there are a few things that we can do to avoid thread hops etc. Over time, I think we should be able to get rid of the proxy too. Your idea of making the Send() implementation in Chromium asynchronous is a good idea, as we'll need that regardless of what we do on the webrtc side.

Patch-set: 3
Reviewer: Gerrit User 5686 <5686@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false