Update patch set 9
Patch Set 9:

> Patch Set 8:
> 
> (1 comment)
> 
> > Patch Set 8: Code-Review+1
> > 
> > (1 comment)
> > 
> > Looks good so far (see comment though). Also let me know if you plan any future work in this area so we don't end up working on the same thing.
> 
> From the email thread my understanding is that you're looking into async Send() on the blink side, correct? I think that's going to be useful and I'm not looking into that. There are a few API related things that I'm fixing for DataChannel (e.g. remove how stats are gathered from DataChannels + exposing the sctp_data_channels_ variable from the controller) and then the rest is basically outlined in this CL.

Ah, I was actually thinking of making it asynchronous on the WebRTC side. I was thinking of the end goal where the signaling and main threads are merged, in which case only two threads would be involved, not three. And given that the lock would only be around adding/removing packets from the queue, I don't imagine it would cause much contention.

But doing it on the Blink side is definitely better in the interim.

However, if *only* Send is called on the network thread, we'd end up with races between the signaling/network threads. Meaning we'll need to either move everything else to the network thread too, or use locks. The former I imagine being easier.

Patch-set: 9