Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> (8 comments)
> 
> Good stuff! some comments

Thanks! and one more thing up to discussion.

This new controller actually decides on bitrate too, we need to add a comment somewhere. But related to that, it might be also ok to use the old bitrate controller, which essentially does the same thing, but seems it has another parameter, which can account for potential overhead mismatch (maybe not a problem anymore, since the overhead is more accurate in current implementation).

Patch-set: 8