Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 7:
> > 
> > (8 comments)
> > 
> > Good stuff! some comments
> 
> Thanks! and one more thing up to discussion.
> 
> This new controller actually decides on bitrate too, we need to add a comment somewhere. But related to that, it might be also ok to use the old bitrate controller, which essentially does the same thing, but seems it has another parameter, which can account for potential overhead mismatch (maybe not a problem anymore, since the overhead is more accurate in current implementation).

The design principle of ANA is that each controller decides one encoding parameters, since making a decision on two or higher-dimensional space is (in principle) hard. I hope not to break this design principle. So it seems that it is essentially the same if we add bitrate controller after the new frame length controller.

Patch-set: 8