Update patch set 7
Patch Set 7:

its a integration test with pre-recorded data which did its job of catching a change in behaviour ;-)

Without this change audio_coding/testfile32kHz encodes as 50 packets kAudioFrameSpeech.
With this change ~50% get classified as kAudioFrameSpeech and the other half as kAudioFrameCN.
According to the comments in TestVadDtx::Run the first second of the file is silence and the test advances half a second to get a mix of silence and speech so the new result makes sense.

The tests should pass now but please check the changes to them make sense.

Patch-set: 7