Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Thoughts? I'm torn, on one hand this saves bitrate (even more so when potentially doing more redundancy) and is consistent with what opus does, on the other hand the extra bitrate is something that I consider affordable.
> 
> Yeah, we're not saving many bits on the wire.

To add some numbers: with this enabled I can go down from 60kbps payload bitrate with to ~40 on average (instead of 30kbps normally used for opus w/o red) and only peak at 60kbps.

That is without dtx (which I have in general only seen in conferencing scenarios). With dtx the bitrate and relatively silent users the average bitrate rate (for the duration of the call) drops considerably anyway and still does drop with DTX.

30kbps more for audio compared to 1mbps+ of video...

Patch-set: 1