Removes locking in TransportFeedbackProxy.
The lock in TransportFeedbackProxy could cause a dead-lock if audio is
included in transport feedback messages, and necessitated a revert:
https://webrtc-review.googlesource.com/c/src/+/178100

This CL removes the lock, and clarifies the assumptions instead:

SetTransportFeedbackObserver() must only ever be called with a single
single instance of TransportFeedbackObserver. This is first done on
creation of an AudioSendStream.

ChannelSend::[De]RegisterSenderCongestionControlObjects() is always
called on the worker thread, and so is
TransportFeedbackProxy::OnTransportFeedback(), since it's call via
AudioSendStream::DeliverRtcp().

Thus the only risk in accessing
TransportFeedbackProxy::transport_feedback_observer_ comes from
TransportFeedbackProxy::OnAddPacket(). Thankfully this is also not an
issue because transport_feedback_ovserver_ is populated before caling
PacketRouter::AddSendRtpModule(), and is then not changed.

For further context, see also internal bug b/153893626

Bug: webrtc:10809
Change-Id: I79b08d8d0e587a59736b383c3596a26836c33d2e