Update patch set 7
Patch Set 7:

(3 comments)

Hi!

I am not quite following along why the feedback observer registration happens so late. Here is my impression of the object hierarchy (skipping unique_ptr-ness when lifetime is shared with owner):

  AudioSendStream {
    RtpTransportControllerSendInterface * const
    ChannelSend {
      TransportFeedbackProxy {
        TransportFeedbackObserver*
      }
    }
  }


The RtpTransportControllerSendInterface is the TransportFeedbackObserver that will be registered within the TransportFeedbackProxy. Apart from this registration, all other creations and references are set up at construction time of the AudioSendStream.

Can the RtpTransportControllerSendInterface (as a pointer to TransportFeedbackObserver) be passed along the constructors to the TransportFeedbackProxy, instead of waiting until a later point (AudioSendStream::ConfigureStream) where threads, locks, and timing considerations make things complicated?
Or do we lose some important conceptual decoupling from doing this?
Or am I wrong, and missing something?

-------------- reasoning -------------

TransportFeedbackProxy is owned by, and shares its lifetime with, ChannelSend. [1]

ChannelSend is owned by, and shares its lifetime with, AudioSendStream. [2]

This CL is about guaranteeing the existence of the TransportFeedbackObserver in the TransportFeedbackProxy in ChannelSend (at the time of the OnAddPacket/OnTransportFeedback calls). This observer object comes from the RtpTransportControllerSendInterface parameter of ChannelSend::RegisterSenderCongestionControlObjects [3]. To be precise, the observer object _is_ the RtpTransportControllerSendInterface [4].

ChannelSend::RegisterSenderCongestionControlObjects is called by the AudioSendStream that owns ChannelSend [5], and the RtpTransportControllerSendInterface argument is a const * member of the AudioSendStream (rtp_transport_), set up in the AudioSendStream constructor [6].

Since both the ChannelSend and the RtpTransportControllerSendInterface are set up from the AudioSendStream constructor, can't we just pass the feedbackobserver (rtp_transport_->transport_feedback_observer()) along to the ChannelSend constructor, instead of waiting until AudioSendStream::ConfigureStream before setting up the feedback observer?

-------------- code links -------------

[1] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/audio/channel_send.cc;l=501;drc=730040c92ecd7b32a67faa689133d3074e149157?originalUrl=https:%2F%2Fcs.chromium.org%2F
[2] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/audio/audio_send_stream.cc;l=119;drc=730040c92ecd7b32a67faa689133d3074e149157?originalUrl=https:%2F%2Fcs.chromium.org%2F
[3] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/audio/channel_send.cc;l=746;drc=730040c92ecd7b32a67faa689133d3074e149157?originalUrl=https:%2F%2Fcs.chromium.org%2F
[4] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/call/rtp_transport_controller_send.cc;l=217;drc=730040c92ecd7b32a67faa689133d3074e149157?originalUrl=https:%2F%2Fcs.chromium.org%2F
[5] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/audio/audio_send_stream.cc;l=304;drc=730040c92ecd7b32a67faa689133d3074e149157?originalUrl=https:%2F%2Fcs.chromium.org%2F
[6] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/audio/audio_send_stream.cc;l=158;drc=730040c92ecd7b32a67faa689133d3074e149157?originalUrl=https:%2F%2Fcs.chromium.org%2F

Patch-set: 7